Changes Styles
Changed the coloring, I think that the only CSS used is the freelancer.mon.css and the bootstrap.css/bootstrapreboot.css. It is Going for a Space theme pulling from both Nasa and SPaceX coloring insperation. Why? Corny reason is you gotta reach for the stars and that is what I want to do OR I am the star people should be going after? Etiher way I am starting to like the basic structure at least.